don't know why the accidental refactoring wasn't included in the last commit, but here it is
